ENH: Placement of text annotations works now interactive.
svn-url: http://viewvc.slicer.org/viewvc.cgi/Slicer4?view=revision&revision=14392
git-svn-id: http://svn.slicer.org/Slicer4/trunk@14392 3bd1e089-480b-0410-8dfb-8563597acbee

+++ QTModules/Annotation/vtkMRMLAnnotationDisplayableManager.h
@@ -73,7 +73,7 @@ protected:
 
 
   /// Callback for click in RenderWindow
-  virtual void OnClickInThreeDRenderWindow(int x, int y);
+  virtual void OnClickInThreeDRenderWindow(float x, float y);
   /// Create a widget.
   virtual vtkAbstractWidget * CreateWidget(vtkMRMLAnnotationNode* node);
   /// Propagate properties of MRML node to widgets.
