+++ QTModules/Annotation/vtkMRMLAnnotationTextDisplayableManager.h
@@ -44,7 +44,7 @@ protected:
   virtual ~vtkMRMLAnnotationTextDisplayableManager(){}
 
   /// Callback for click in RenderWindow
-  virtual void OnClickInThreeDRenderWindow(int x, int y);
+  virtual void OnClickInThreeDRenderWindow(float x, float y);
   /// Create a widget.
   virtual vtkAbstractWidget * CreateWidget(vtkMRMLAnnotationNode* node);
   /// Propagate properties of MRML node to widgets.
